actually restoring nt
