+++ src/tj2_networktables/CMakeLists.txt
@@ -118,7 +118,7 @@ catkin_python_setup()
 ## DEPENDS: system dependencies of this project that dependent projects also need
 catkin_package(
   INCLUDE_DIRS include
-  LIBRARIES tj2_networktables diffyjr_networktables robot_2022_networktables scorpion_networktables
+  LIBRARIES tj2_networktables diffyjr_networktables scorpion_networktables
   CATKIN_DEPENDS 
     rospy
     std_msgs
@@ -150,10 +150,6 @@ add_library(tj2_networktables
 
 add_library(diffyjr_networktables
   src/diffyjr_networktables.cpp
-)
-
-add_library(robot_2022_networktables
-  src/robot_2022_networktables.cpp
 )
 
 add_library(scorpion_networktables
@@ -190,11 +186,6 @@ target_link_libraries(diffyjr_networktables
     ntcore wpiutil tj2_networktables
 )
 
-target_link_libraries(robot_2022_networktables
-    ${catkin_LIBRARIES}
-    ntcore wpiutil tj2_networktables
-)
-
 target_link_libraries(scorpion_networktables
     ${catkin_LIBRARIES}
     ntcore wpiutil tj2_networktables
@@ -202,14 +193,13 @@ target_link_libraries(scorpion_networktables
 
 target_link_libraries(tj2_networktables_node
     ${catkin_LIBRARIES}
-    ntcore wpiutil tj2_networktables robot_2022_networktables diffyjr_networktables scorpion_networktables
+    ntcore wpiutil tj2_networktables diffyjr_networktables scorpion_networktables
 )
 
 ## Add cmake target dependencies of the executable
 ## same as for the library above
 add_dependencies(tj2_networktables ${tj2_networktables_EXPORTED_TARGETS} ${catkin_EXPORTED_TARGETS})
 add_dependencies(diffyjr_networktables ${diffyjr_networktables_EXPORTED_TARGETS} ${catkin_EXPORTED_TARGETS})
-add_dependencies(robot_2022_networktables ${robot_2022_networktables_EXPORTED_TARGETS} ${catkin_EXPORTED_TARGETS})
 add_dependencies(scorpion_networktables ${scorpion_networktables_EXPORTED_TARGETS} ${catkin_EXPORTED_TARGETS})
 add_dependencies(tj2_networktables_node ${tj2_networktables_node_EXPORTED_TARGETS} ${catkin_EXPORTED_TARGETS})
