+++ src/tj2_networktables/src/tj2_networktables_node.cpp
@@ -1,5 +1,4 @@
 #include "diffyjr_networktables.h"
-#include "robot_2022_networktables.h"
 #include "scorpion_networktables.h"
 
 int main(int argc, char **argv)
@@ -13,9 +12,6 @@ int main(int argc, char **argv)
     TJ2NetworkTables* nt_interface = NULL;
     if (nt_interface_name.compare("diffyjr") == 0) {
         nt_interface = new DiffyJrNetworkTables(&nh);
-    }
-    else if (nt_interface_name.compare("robot_2022") == 0) {
-        nt_interface = new Robot2022NetworkTables(&nh);
     }
     else if (nt_interface_name.compare("scorpion") == 0) {
         nt_interface = new ScorpionNetworkTables(&nh);
